[FIX] Zero storage operations' result arguments
Write/read/erase shall zero the result arguments at the beginning, so that in
case of a failure, a bug on the user side of not checking the error
code is avoided.

Even if the error code is not checked by the user, he will still get 0 number
of process bytes as a result.